Allow to fingerprint all pickleable objects.
Turns out we need to be able to fingerprint functions as some decoders
use them as attributes. Instead of relying on our own code, it seems
better to rely on the pickle module. Everything which can be pickled
should be ok to fingerprint and everything else shouldn't. Also,
simplifies the code a lot.