Made BasalGanglia and Thalamus functions
Instead of subclasses. Took a bit of creative rearranging,
but I think the end result is a fair bit simpler than what
we had before.

Notably, I used the fact that there's a separate `Config` stack
for configuring certain things in the BG instead of using kwargs.
This means that if you want to change something about all of the
AMPA connections, you don't have to edit the BG at all, you just
have to edit `nengo.networks.actionselection.ampa_config`,
which you can do in a model script.

Also of note is the pattern of accepting a premade network, or
making a new one if none is passed. This has a few concrete benefits.
One is that the BG / Thalamus doesn't need to accept label, seed,
and add_to_container arguments because if you want something other
than the default, you just quickly make a network with those values
and pass it in (this is done in `test_actionselection.py`).
Similarly, if you want to change something like the max firing rates,
you can make a new network, change the max firing rates in that
network's config, and it will just use them.

Other changes:
- Made it easier to make `Config` objects by allowing the user
  to specify the objects that can be configured by passing them
  to `__init__`.