Converted more networks to non-subclasses
Specifically, Integrator, Oscillator, Product, and
the working memory networks. With the patterns described in
the previous commit, was able to make a lot of the functions
slightly simpler by using the config system. `test_product`
provides a good example of passing in a network preconfigures
(in this case, to use a particular neuron type).

Other misc changes:
- Fixed a small bug in `Config.default`, when using `Config`
  instances that don't configure the class you want.
- Made the `Product` easier to use (hopefully) by asking users
  to specify the input radius; then we just square it (with some
  extra fudge factor). Before, it was being multiplied by sqrt(2),
  which only sort of makes sense to me.
- Had to up the tolerances on the Product test because the absolute
  values are higher now (got rid of my sqrt(radius) fix from a few
  commits back).