Changed from global configs to arguments in BG
I think the resulting code is better, though there are two places
where it's a bit shady. The first is that, in order to allow the user
to overwrite the BG's config (which sets the radius, encoders, and
decoder solver in the network), I have the passed in network take
config precedence over the BG's config. While this is, I think, the
most flexible thing to do, and mostly safe because in most cases the
user will want to keep the BG in an isolated network, it is possible
that people might pass in an existing network with the radius or
encoders set, causing subtle bugs. We could warn in this case, but
it'd be a bit annoying if the user is intentionally changing these
things.

The second is that if the user passes in a config object for the
AMPA or GABA connections, we raise a warning if they don't set the
synapse. In this case, I think this is the correct behavior because
the AMPA and GABA configs only set the synapse; if you're overwriting
them, it's very likely to be to play around the synapse model.
On the other hand, if you just want to change the decoder solver or
something, we at least do the "right thing" by setting the synapse,
but we do this by just setting the synapse in the passed config,
which is a bit heavy handed.

Other changes:
- Implemented __contains__ to make it easy to see if you have set
  a value for a particular config (e.g., to see if the radius has
  been set, you can do `'radius' in config[nengo.Ensemble]`).
- Added the `nested` function to `nengo.utils.compat`. This is the
  only way we can do nested context managers in a way that works
  on Python 2.6 and above; if we don't support Python 2.6 then we
  don't have to include this function.