PEP8, Python3 fixes for cache branch
Most of the changes were related to strings, which is where the
major Python 2 / 3 differences are. Since pickle serialization uses
strings, we have to encode them in Python 3, but no Python 2;
and, in another case, encode in Python 2 but not Python 3.
In any case, all of the test cases pass in both Python 2 and 3 now.

The other changes are mostly cosmetic, or PEP8 related.

Oh, and I also removed the seed from `Model`, as it looks like at some
point the top-level network seed was used to determine if caching
should be done. At least, I wasn't able to see any uses of model.seed.